# AutoPulse-backend/common/src/main/java/com/autopulse/common/api/ErrorResponse.java
package com.autopulse.common.api;

import java.time.Instant;
import java.util.Map;

public record ErrorResponse(
        Instant timestamp,
        int status,
        String error,
        String message,
        String path,
        String correlationId,
        String code,
        Map<String, Object> details
) {}