# AutoPulse-backend/fleet-service/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: AutoPulse Fleet Service API
  version: 1.0.0
  description: |
    OpenAPI contract for the fleet-service microservice
    This API manages depots, vehicles, vehicle assignments, vehicle documents,
    and vehicle maintenance records.
servers:
  - url: http://localhost:8083
    description: fleet-service
  - url: http://localhost:8080
    description: API Gateway
...
security:
  - bearerAuth: []
paths:
...
  /api/fleet/vehicle-assignments:
    post:
      tags: [Vehicle Assignments]
      summary: Assign vehicle to courier
      operationId: createVehicleAssignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleAssignmentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleAssignmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [Vehicle Assignments]
      summary: List vehicle assignments
      operationId: getVehicleAssignments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVehicleAssignmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/fleet/vehicle-assignments/{id}:
    get:
      tags: [Vehicle Assignments]
      summary: Get vehicle assignment by id
      operationId: getVehicleAssignmentById
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleAssignmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/fleet/vehicle-assignments/{id}/end:
    patch:
      tags: [Vehicle Assignments]
      summary: End vehicle assignment
      operationId: endVehicleAssignment
      parameters:
        - $ref: '#/components/parameters/AssignmentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleAssignmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/fleet/couriers/{courierId}/vehicle-assignment:
    get:
      tags: [Vehicle Assignments]
      summary: Get active vehicle assignment for courier
      operationId: getActiveVehicleAssignmentByCourier
      parameters:
        - $ref: '#/components/parameters/CourierId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleAssignmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/fleet/couriers/{courierId}/vehicle-assignments:
    get:
      tags: [Vehicle Assignments]
      summary: List assignments by courier
      operationId: getVehicleAssignmentsByCourier
      parameters:
        - $ref: '#/components/parameters/CourierId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVehicleAssignmentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/fleet/vehicles/{vehicleId}/documents:
    post:
      tags: [Vehicle Documents]
      summary: Create vehicle document
      operationId: createVehicleDocument
      parameters:
        - $ref: '#/components/parameters/VehicleIdPathVehicleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVehicleDocumentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Vehicle Documents]
      summary: List documents by vehicle
      operationId: getVehicleDocumentsByVehicle
      parameters:
        - $ref: '#/components/parameters/VehicleIdPathVehicleId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVehicleDocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/fleet/documents/expired:
    get:
      tags: [Vehicle Documents]
      summary: List expired vehicle documents
      operationId: getExpiredVehicleDocuments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVehicleDocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/fleet/documents/{documentId}:
    get:
      tags: [Vehicle Documents]
      summary: Get vehicle document by id
      operationId: getVehicleDocumentById
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDocumentResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Vehicle Documents]
      summary: Update vehicle document
      operationId: updateVehicleDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehicleDocumentRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDocumentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Vehicle Documents]
      summary: Delete vehicle document
      operationId: deleteVehicleDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
      description: Page size
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: name,asc
      description: Sort format is property,direction. Example name,asc
...
    VehicleIdPathVehicleId:
      name: vehicleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AssignmentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CourierId:
      name: courierId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    VehicleAssignmentStatus:
      type: string
      enum: [ACTIVE, ENDED]
    VehicleDocumentType:
      type: string
      enum: [REGISTRATION_CERTIFICATE, INSURANCE, INSPECTION, LEASE_CONTRACT, OTHER]
...
    CreateVehicleAssignmentRequest:
      type: object
      required: [vehicleId, courierId]
      properties:
        vehicleId:
          type: string
          format: uuid
        courierId:
          type: string
          format: uuid
        startDate:
          type: string
          format: date-time
          nullable: true
    VehicleAssignmentResponse:
      type: object
      required: [id, vehicleId, courierId, status]
      properties:
        id:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        courierId:
          type: string
          format: uuid
        vehicleLicensePlate:
          type: string
        status:
          $ref: '#/components/schemas/VehicleAssignmentStatus'
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
    CreateVehicleDocumentRequest:
      type: object
      required: [type, documentNumber, issuedAt, expiresAt]
      properties:
        type:
          $ref: '#/components/schemas/VehicleDocumentType'
        documentNumber:
          type: string
          minLength: 1
          maxLength: 100
        issuedAt:
          type: string
          format: date
        expiresAt:
          type: string
          format: date
        fileUrl:
          type: string
          format: uri
          nullable: true
    UpdateVehicleDocumentRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/VehicleDocumentType'
        documentNumber:
          type: string
          minLength: 1
          maxLength: 100
        issuedAt:
          type: string
          format: date
        expiresAt:
          type: string
          format: date
        fileUrl:
          type: string
          format: uri
          nullable: true
    VehicleDocumentResponse:
      type: object
      required: [id, vehicleId, type, documentNumber, issuedAt, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        vehicleId:
          type: string
          format: uuid
        type:
          $ref: '#/components/schemas/VehicleDocumentType'
        documentNumber:
          type: string
        issuedAt:
          type: string
          format: date
        expiresAt:
          type: string
          format: date
        expired:
          type: boolean
        fileUrl:
          type: string
          format: uri
          nullable: true
...
    PageVehicleAssignmentResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/VehicleAssignmentResponse'
    PageVehicleDocumentResponse:
      allOf:
        - $ref: '#/components/schemas/PageResponse'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/VehicleDocumentResponse'
...
    PageResponse:
      type: object
      required: [content, totalElements, totalPages, size, number]
      properties:
        content:
          type: array
          items:
            type: object
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
        empty:
          type: boolean
    ErrorResponse:
      type: object
      required: [timestamp, status, errorCode, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        errorCode:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        path:
          type: string
          example: /api/fleet/vehicles
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: licensePlate
        message:
          type: string
          example: must not be blank